backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
langchain>=0.0.267
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
pytest>=7.0.0
